Ajax/Json categorias

+++ carro.php
@@ -14,17 +14,39 @@
 
 
   $action = "view";
-  if(isset($_GET['action'])){
-    $action = $_GET['action'];
-  }
-  if($action == "add"){
-    $carrito->addItem($_GET["id"], $_GET['cantidad']);
-  }
-  if($action == "delete"){
-    $carrito->deleteItem($_GET["id"]);
-  }
-  if($action == "empty"){
-    $carrito->empty();
+  if (($_SERVER["REQUEST_METHOD"] == "POST")){
+    //Por javascript siempre llamamos con método POST
+    if (isset($_POST["action"])){
+        $action = $_POST["action"];
+    }
+    if ($action == "add"){
+      //Pero accedo al 'id' mediante GET (ya que está en la url)
+      if (!$carrito->itemExists($_GET["id"])){
+        //Cuando es add, siempre añadimos 1 pero sólo si no estaba ya antes en el carro
+        $carrito->addItem($_GET["id"], 1);
+      }
+      //Siempre devolvemos en itemCount los que ya hay ahora en el carro de ese producto
+      echo json_encode(array("HOME"=> "./", "itemCount"=>$carrito->getItemCount($_GET["id"]), "cuantos"=> $carrito->howMany(), "total"=> $carrito->precioFinal()));
+      exit();
+    }elseif ($action == "update"){
+        //Cuando es update (desde el botón de actualizar de la ventana modal, la cantidad es la que introduce el usuario)
+        $carrito->addItem($_GET["id"], $_POST["cantidad"]);
+        echo json_encode(array("HOME"=> "./", "itemCount"=>$carrito->getItemCount($_GET["id"]), "cuantos"=> $carrito->howMany(), "total"=> $carrito->precioFinal()));
+        exit();
+    }
+  }else{
+    if(isset($_GET['action'])){
+      $action = $_GET['action'];
+    }
+    if($action == "add"){
+      $carrito->addItem($_GET["id"], $_GET['cantidad']);
+    }
+    if($action == "delete"){
+      $carrito->deleteItem($_GET["id"]);
+    }
+    if($action == "empty"){
+      $carrito->empty();
+    }
   }
   include("./include/header.php");
 ?>
